Changes to UserActivity placement
